Create Side by Side (Amateaur Web Site)
This is one of my first projects. It's a project for mi High School. Is a web site for a fast food restaurant to my zone. I used a template and I adapted it for my project. It Don't has a good responsive, and I don't know how fix it. Any correction that you can give me will be welcome. Thanks :)... Sorry fro my english, i'm learning yet.